Add files via upload
Title: Experience in Creating Chatbots Using BERT, GPT-2, Neural Networks, and TF-IDF

Description:
I recently had the exciting opportunity to collaborate with my peers, Amine Belkessa and Zeraoui Yani, in creating two chatbots that engage in conversations with each other. Our project was an in-depth exploration of various methods, including the use of BERT, GPT-2, neural networks, and TF-IDF.

We encountered several challenges throughout our journey. One of the main hurdles was the scarcity of relevant code for chatbots, especially those not intended for sale. We had to think creatively and resourcefully to adapt the available resources to our specific needs.

Another major challenge was finding a high-quality dataset organized in a question-answer (Q/A) format. Locating a dataset suitable for our precise requirements required significant time and effort, but it was worthwhile to ensure the quality and relevance of our chatbots.

Additionally, hardware constraints also posed challenges, sometimes slowing down our progress. Despite these obstacles, we persevered and were able to create functional chatbots capable of engaging in coherent and meaningful conversations.

This project was a valuable experience that deepened our understanding of natural language processing technologies and honed our skills in chatbot design and development. We are proud of the work accomplished and excited about future possibilities in this constantly evolving field.